refactor(service/html): add interface
