+++ src/test/java/com/jiuzhang/seckill/service/ActivityHtmlPageServiceTest.java
@@ -2,7 +2,7 @@ package com.jiuzhang.seckill.service;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-import com.jiuzhang.flashsale.service.ActivityHtmlPageService;
+import com.jiuzhang.flashsale.service.impl.ActivityHtmlPageServiceImpl;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,7 +12,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 class ActivityHtmlPageServiceTest {
 
     @Autowired
-    private ActivityHtmlPageService activityHtmlPageService;
+    private ActivityHtmlPageServiceImpl activityHtmlPageService;
 
     @Test
     void testCreateActivityHtml() {
